+++ Classes/Csv/CsvDownloader.php
@@ -19,34 +19,9 @@ class CsvDownloader
     use TranslateTrait;
 
     /**
-     * @var int
-     */
-    public const CSV_TYPE_NUMBER = 736;
-
-    /**
-     * @var int HTTP status code for "page not found"
-     */
-    private const NOT_FOUND = 404;
-
-    /**
-     * @var int HTTP status code for "access denied"
-     */
-    private const ACCESS_DENIED = 403;
-
-    /**
      * @var Configuration
      */
     protected $configuration;
-
-    /**
-     * @var string the TYPO3 mode set for testing purposes
-     */
-    private $typo3Mode = '';
-
-    /**
-     * @var int the HTTP status code of error
-     */
-    private $errorType = 0;
 
     public function __construct()
     {
@@ -60,20 +35,18 @@ class CsvDownloader
      */
     public function main(): string
     {
-        try {
-            switch ((string)GeneralUtility::_GET('table')) {
-                case 'tx_seminars_seminars':
-                    $result = $this->createAndOutputListOfEvents((int)GeneralUtility::_GET('pid'));
-                    break;
-                case 'tx_seminars_attendances':
-                    $result = $this->createAndOutputListOfRegistrations((int)GeneralUtility::_GET('eventUid'));
-                    break;
-                default:
-                    $result = $this->addErrorHeaderAndReturnMessage(self::NOT_FOUND);
-            }
-        } catch (\Exception $exception) {
-            HeaderProxyFactory::getInstance()->getHeaderProxy()->addHeader('Status: 500 Internal Server Error');
-            $result = $exception->getMessage() . "\n\n" . $exception->getTraceAsString() . "\n\n";
+        switch ((string)GeneralUtility::_GET('table')) {
+            case 'tx_seminars_seminars':
+                $result = $this->createAndOutputListOfEvents((int)GeneralUtility::_GET('pid'));
+                break;
+            case 'tx_seminars_attendances':
+                $result = $this->createAndOutputListOfRegistrations((int)GeneralUtility::_GET('eventUid'));
+                break;
+            default:
+                throw new \InvalidArgumentException(
+                    'The parameter "table" must be set to either "tx_seminars_seminars" or "tx_seminars_attendances".',
+                    1671155057
+                );
         }
 
         return $result;
@@ -94,18 +67,7 @@ class CsvDownloader
     {
         $listView = GeneralUtility::makeInstance(DownloadRegistrationListView::class);
 
-        $pageUid = (int)GeneralUtility::_GET('pid');
-        if ($eventUid > 0) {
-            if (!$this->hasAccessToEventAndItsRegistrations($eventUid)) {
-                return $this->addErrorHeaderAndReturnMessage($this->errorType);
-            }
-            $listView->setEventUid($eventUid);
-        } else {
-            if (!$this->canAccessRegistrationsOnPage($pageUid)) {
-                return $this->addErrorHeaderAndReturnMessage(self::ACCESS_DENIED);
-            }
-            $listView->setPageUid($pageUid);
-        }
+        $listView->setEventUid($eventUid);
 
         $this->setContentTypeForRegistrationLists();
 
@@ -149,7 +111,7 @@ class CsvDownloader
     public function createAndOutputListOfEvents(int $pageUid): string
     {
         if ($pageUid <= 0) {
-            return $this->addErrorHeaderAndReturnMessage(self::NOT_FOUND);
+            throw new \InvalidArgumentException('The parameter $pageUid must be > 0.', 1671155090);
         }
 
         $this->setContentTypeForEventLists();
@@ -172,36 +134,6 @@ class CsvDownloader
         $eventListView->setPageUid($pageUid);
 
         return $eventListView->render();
-    }
-
-    /**
-     * Checks whether the list of registrations is accessible, ie.
-     * 1. CSV access is allowed for testing purposes, or
-     * 2. the logged-in BE user has read access to the registrations table and
-     *    read access to *all* pages where the registration records of the
-     *    selected event are stored.
-     *
-     * @param int $eventUid UID of the event record for which access should be checked, must be > 0
-     *
-     * @return bool true if the list of registrations may be exported as CSV
-     */
-    protected function canAccessListOfRegistrations(int $eventUid): bool
-    {
-        switch ($this->getTypo3Mode()) {
-            case 'BE':
-                $result = true;
-                break;
-            case 'FE':
-                $event = GeneralUtility::makeInstance(LegacyEvent::class, $eventUid, false, true);
-                GeneralUtility::makeInstance(FrontEndRegistrationAccessCheck::class)->setEvent($event);
-
-                $result = GeneralUtility::makeInstance(FrontEndRegistrationAccessCheck::class)->hasAccess();
-                break;
-            default:
-                $result = false;
-        }
-
-        return $result;
     }
 
     /**
@@ -232,113 +164,5 @@ class CsvDownloader
         $headerProxy = HeaderProxyFactory::getInstance()->getHeaderProxy();
         $headerProxy->addHeader('Content-type: text/csv; header=present; charset=utf-8');
         $headerProxy->addHeader('Content-disposition: attachment; filename=' . $csvFileName);
-    }
-
-    /**
-     * Adds a status header and returns an error message.
-     *
-     * @param int $errorCode
-     *        the type of error message, must be ACCESS_DENIED or NOT_FOUND
-     *
-     * @return string the error message belonging to the error code, will not be empty
-     *
-     * @throws \InvalidArgumentException
-     */
-    private function addErrorHeaderAndReturnMessage(int $errorCode): string
-    {
-        switch ($errorCode) {
-            case self::ACCESS_DENIED:
-                HeaderProxyFactory::getInstance()->getHeaderProxy()->addHeader('Status: 403 Forbidden');
-                $result = $this->translate('message_403');
-                break;
-            case self::NOT_FOUND:
-                HeaderProxyFactory::getInstance()->getHeaderProxy()->addHeader('Status: 404 Not Found');
-                $result = $this->translate('message_404');
-                break;
-            default:
-                throw new \InvalidArgumentException('"' . $errorCode . '" is no legal error code.', 1333292523);
-        }
-
-        return $result;
-    }
-
-    /**
-     * Checks whether the currently logged-in BE-User is allowed to access the registrations records on the given page.
-     *
-     * @param int $pageUid PID of the page to check the access for, must be >= 0
-     *
-     * @return bool
-     *         TRUE if the currently logged-in BE-User is allowed to access the registrations records,
-     *         FALSE if the user has no access or this function is called in FE mode
-     */
-    private function canAccessRegistrationsOnPage(int $pageUid): bool
-    {
-        switch ($this->getTypo3Mode()) {
-            case 'BE':
-                $result = true;
-                break;
-            case 'FE':
-                // The fall-through is intentional.
-            default:
-                $result = false;
-        }
-
-        return $result;
-    }
-
-    /**
-     * Returns the mode currently set in TYPO3_MODE.
-     *
-     * @return string either "FE" or "BE" representing the TYPO3 mode
-     */
-    private function getTypo3Mode(): string
-    {
-        if ($this->typo3Mode !== '') {
-            return $this->typo3Mode;
-        }
-
-        return TYPO3_MODE;
-    }
-
-    /**
-     * Sets the TYPO3_MODE.
-     *
-     * The value is stored in the member variable $this->typo3Mode
-     *
-     * This function is for testing purposes only!
-     *
-     * @param string $typo3Mode the TYPO3_MODE to set, must be "BE" or "FE"
-     */
-    public function setTypo3Mode(string $typo3Mode): void
-    {
-        $this->typo3Mode = $typo3Mode;
-    }
-
-    /**
-     * Checks whether the currently logged in BE-User has access to the given
-     * event and its registrations.
-     *
-     * Stores the type of the error in $this->errorType
-     *
-     * @param int $eventUid the event to check the access for, must be >= 0 but not necessarily point
-     *        to an existing event
-     *
-     * @return bool true if the event record exists and the BE User has
-     *                 access to the registrations belonging to the event,
-     *                 false otherwise
-     */
-    private function hasAccessToEventAndItsRegistrations(int $eventUid): bool
-    {
-        $result = false;
-
-        if (LegacyEvent::fromUid($eventUid, true) === null) {
-            $this->errorType = self::NOT_FOUND;
-        } elseif ($this->canAccessListOfRegistrations($eventUid)) {
-            $result = true;
-        } else {
-            $this->errorType = self::ACCESS_DENIED;
-        }
-
-        return $result;
     }
 }