+++ Tests/Unit/Configuration/MyVipEventsConfigurationCheckTest.php
@@ -49,16 +49,4 @@ final class MyVipEventsConfigurationCheckTest extends UnitTestCase
         $result = $this->subject->getWarningsAsHtml();
         self::assertNotSame([], $result);
     }
-
-    /**
-     * @test
-     */
-    public function checkWithEmptyConfigurationUsesProvidedNamespaceForErrors(): void
-    {
-        $this->subject->check();
-
-        $result = $this->subject->getWarningsAsHtml();
-        self::assertArrayHasKey(1, $result);
-        self::assertStringContainsString('plugin.tx_seminars_pi1', $result[1]);
-    }
 }